가격정보 int 에서 double 로 변경테스트

+++ Sources/DealiDesignKit/Components/TextField/TextInputs/DealiTextInput_v2.swift
@@ -221,9 +221,9 @@ public final class DealiTextInput_v2: UIView, DealiTextField {
     }
     
     /// 최소 금액 (type이 price일경우 사용)
-    public var minPrice = 0
+    public var minPrice = 0.0
     /// 최대 금액 (type이 price일경우 사용)
-    public var maxPrice = 10000000
+    public var maxPrice = 10000000.0
     /// 최소 최대 금액 밖의 가격은 입력이 안되도록
     public var blockOutOfRangePriceInput: Bool = true
     
@@ -272,7 +272,7 @@ public final class DealiTextInput_v2: UIView, DealiTextField {
             .distinctUntilChanged()
             .bind(with: self) { owner, text in
                 if owner.textInputFormat == .price,
-                   let price = Int(text.replacingOccurrences(of: ",", with: "")) {
+                   let price = Double(text.replacingOccurrences(of: ",", with: "")) {
                     owner.textField.text = price.commaString()
                 }
             }
@@ -351,7 +351,7 @@ extension DealiTextInput_v2: UITextFieldDelegate {
             currentPriceText += string
         }
         
-        guard let price = Int(currentPriceText) else { return isNumber || string.isEmpty }
+        guard let price = Double(currentPriceText) else { return isNumber || string.isEmpty }
         
         return price <= maxPrice && price >= minPrice
     }
